Implement PaperTrail for auditing
https://eaflood.atlassian.net/browse/RUBY-203

PaperTrail is a gem used for keeping a record of when and how a record is changed. We use it to keep track of changes to registrations. This information will then be shown in the WEX back office.

This PR will install and configure PaperTrail so we can migrate our existing audit history and continue to build on it.