Fix the intermittent failure of the ExemptionsForm spec (#87)
This fixes an intermittent failure in the ExemptionsForm spec. The spec was expecting to compare sorted arrays of exemption codes. Since the code is a combination of "U" and an integer, the sorting was insufficient when the integer was greater than 9 since string sorting will sort "U10" before "U9". Using `match_array` fixes this as the order is not relevant.