Add PaperTrail to models and run migration
To match the configuration of the old engine, we need to track versions of the Registration and the Address (and back office users, but that's not for here).

This commit adds and runs the migration for the new `versions` table in the database, and adds PaperTrail to the models we want to track.