Implement PaperTrail for auditing (#86)
* Implement PaperTrail for auditing

https://eaflood.atlassian.net/browse/RUBY-203

PaperTrail is a gem used for keeping a record of when and how a record is changed. We use it to keep track of changes to registrations. This information will then be shown in the WEX back office in a future story.

This PR installs and configure PaperTrail so we can migrate our existing audit history and continue to build on it. It adds and runs the migration for the new `versions` table in the database, and adds PaperTrail to the models we want to track.

To match the configuration of the old engine, we need to track versions of the Registration and the Address. There are certain Address attributes which we don't track - when these attributes are modified, a new version should not be created unless another attribute is also changed.

PaperTrail also logs a `whodunnit` when new registration or address version is created. The front-office has no current_user, but the back-office does, and we want to track who that was. So we allow the engine config to specify whether to track a current_user or not.

The PaperTrail RSpec helper switches PaperTrail off by default when running tests (which should improve our speed) but also lets us use some extra matches when tests are enabled (only in the PaperTrail shared example).