Save the assistance mode for a registration (#98)
https://eaflood.atlassian.net/browse/RUBY-200

A registration should have a different `assistance_mode` depending on whether it was submitted through the front-office or back-office. This allows us to track what proportion of registrations are assisted digital.

The host apps can set a default `assistance_mode` in the engine initializer.

This commit also adds some tests for the RegistrationCompleteService, to cover the new feature and some of the preexisting behaviour. These are modelled on the RenewalCompletionService from the waste carriers engine.