Add rake task to generate privacy policy PDF
https://eaflood.atlassian.net/browse/RUBY-216

Data protection asked us to include a copy of the privacy policy for those users who complete their registration via NCCC as assisted digital.

The team looked at the problem and decided it would just be simpler and meet the need more if we just attached a copy of the policy to all emails.

However as the policy contains no dynamic elements there would be no need to generate it each time we send the email. It would be better if we could generate the PDF ahead of time, add it as a static asset to the projects and then just attach that asset to the confirmation email.

There are no doubt online tools that could take html and generate a PDF for us, however we have already built into the engine the ability to generate PDF's from html. So why not just use that funcationality? 🤔💡

Hence this change which adds a html version of the privacy policy, and a rake task to convert that to a pdf file. It outputs to a tmp folder to avoid the file getting accidently commited here. The intention is that you generate the file when needed here, and then simply copy and past it to the apps that host the engine.

As we expect to do this so infrequently, we decided this simple approach was preferred than trying to do anything to clever in the code.

---

WIP