Add rspec test for privacy pdf generator + refactor
In order to test the rake task without actually running rake itself, all the code in the rake task was refactored to class in `lib/`.

This meant we could test the functionality without using Rake. It did also mean though we also had to change some of the internal functionality of the method to no longer rely on Rake methods directly.