Be explicit with link content
This change updates the content slightly so that the links to external web pages are displayed in full, so should a user need to they can see the url to copy.

This has nothing to do with the fact I spotted that links don't seem to work in wicked_pdf generated PDF's 🤬🤬🤬🤬🤬🤬🤬🤬.

The only reference I can see to this issue is https://github.com/mileszs/wicked_pdf/issues/608. When you start following some of the links I can see their [example pdf](https://wicked-pdf-issues.herokuapp.com/pages/issue_608.pdf) has working links. There is also a suggestion in this [stackoverflow post](https://stackoverflow.com/a/38616579) post that you just need to ensure you include the protocol, but we already do.

However I tried various methods including

- using `WickedPdf.new.pdf_from_html_file`
- using `WickedPdf.new.pdf_from_url` and pointing to their own example https://wicked-pdf-issues.herokuapp.com/pages/issue_608
- passing in a string that just represents a link

Nothing seemed to work. The one thing that did was explicit links. So by actually putting the link in the content, my OSX preview and Chrome started to recognise the links i.e. the mouse cursor would change and they became clickable.

It might actually be an underlying issue with wkhtmltopdf according as [this issue comment](https://github.com/wkhtmltopdf/wkhtmltopdf/issues/1692#issuecomment-59231801) is a very close description of my issue and the fix I implemented.

This is _slightly_ frustrating in so far as if I upload the privacy.html file to https://html2pdf.com/ I get a PDF with working links and rendered in the same way. Sort of makes my efforts today meaningless 😩🤦.