Add missing tests for GeneratePdfService
Actually created this as part of [PR 117](https://github.com/DEFRA/waste-exemptions-engine/pull/117) but there is a good chance that is going to be closed without merging.

However there is no sense in chucking away the work of creating the tests hence this standalone change to add the missing tests for the service.