Add html view which is used to generate PDF privacy
https://eaflood.atlassian.net/browse/RUBY-216

Data protection asked us to include a copy of the privacy policy for those users who complete their registration via NCCC as assisted digital.

The team looked at the problem and decided it would just be simpler and meet the need more if we just attached a copy of the policy to all emails.

However as the policy contains no dynamic elements there would be no need to generate it each time we send the email. It would be better if we could generate the PDF ahead of time, add it as a static asset to the projects and then just attach that asset to the confirmation email.

The need to generate and update the PDF will be very infrequent (we are talking years) so rather than try and do anything ourselves we will just take advantage of available online generators.

For them to work they need to be provided with the HTML to convert, hence this change which adds pdf specific version of the privacu policy we can upload to be converted.

N.B. At this time (April 2019) we are using https://html2pdf.com/ to convert the html to PDF.