Attach privacy policy pdf to confirmation
https://eaflood.atlassian.net/browse/RUBY-208

Data protection asked us to include a copy of the privacy policy for those users who complete their registration via NCCC as assisted digital.

The team looked at the problem and decided it would just be simpler and meet the need more if we just attached a copy of the policy to all emails.

However as the policy contains no dynamic elements there would be no need to generate it each time we send the email. It would be better if we could generate the PDF ahead of time, add it as a static asset to the projects and then just attach that asset to the confirmation email.

[PR 119[(https://github.com/DEFRA/waste-exemptions-engine/pull/119) created a html view which is used to generate the privacy policy pdf. This change adds the generated pdf to the project (added to lib as it seems to be the best place for it!) and amends the confirmation mailer code to read the file and add it as an attachment.