EditedRegistraton factory uses a Registration when initialising
When we create a new EditedRegistration, it will need to copy all its data from an existing Registration, similar to how the waste carriers engine currently creates new TransientRegistrations for renewing.

So that we have data to copy, we need a new trait for the Registration which has all the expected data.