Add alternative edit state machine for EditedRegistrations
Boom, double state machines! Both use the same column and the same forms (with a few exceptions which will be added in future commits). However, the allowed states and transitions are determined based on which state machine the model is using.

For EditedRegistrations, this is the new CanChangeEditWorkflowStatus concern. TransientRegistrations currently use the existing new registration journey state machine. Once NewRegistrations are created, I'll be refactoring so this logic about which concern to use isn't so disgusting.

The edit state machine mostly follows a pattern of looping from `edit_form` (the overview edit page) to an individual form, and then when that is submitted returns to `edit_form`. There are a few exceptions (non-England locations, address lookups).

Some values cannot currently be edited in WEX – business type and the exemptions. These states are not included in the state machine as the forms should not be available to edit.