Create NewRegistration
NewRegistration is basically what TransientRegistrations used to be by default – a registration which is in the process of being created.

This commit creates the NewRegistration, but also updates the rest of the app to use it as appropriate. It also includes refactoring to move any logic which should be specific to NewRegistrations, as well as removing some temporary hacky stuff from EditedRegistration now that we no longer have to block it from inheriting some of the NewReg behaviour.