Add `layout: false` when rendering PDF views to string.
This solves a bug on the most recent refactoring for which the `render_to_string` method fails outside of a request scope
because the included application Layout depends on the presence of a `@_request` object. Since we don't need our PDF
to be rendered with a layout, this PR explicitely tells rails not to fetch the view's layout.