Refactor address workflow tests to use shared examples (#135)
https://eaflood.atlassian.net/browse/RUBY-62

This PR adds some shared examples for the address workflow tests.

This will reduce duplication now, and prevent even more of it when we add more workflow tests for the edit workflow.

These new examples differ slightly from the other transitions in that they don't assume the subject of the example is a TransientRegistration. Little bit of future-proofing.