Add edit workflow
This works in the same way as the new registration workflow - we have a central state machine managing everything, which is included in the EditRegistration as a concern.

The key difference is that the edit workflow is much less sequential. Generally, users will go from the edit overview form to a specific form, and then back again. In some cases, editing a value requires going through multiple forms/states before returning to the edit overview.

Because the workflow is so much less linear, we end up with a lot more transitions. However, most of the states involved are exactly the same as the other workflow, with the exception of a few edit-specific states.