Subclass TransientRegistration with NewRegistration (#138)
https://eaflood.atlassian.net/browse/RUBY-62

This is part of a series of PRs to add editing functionality to the back office.

When we first created TransientRegistration, it was only used for the new registration journey. However, now we want to have multiple times of TransientRegistrations for different occasions when we want to modify Registration data.

This PR adds the ability to subclass TransientRegistration, and also creates a NewRegistration to perform the functions specific to the new registration journey.

* Add type to TransientRegistrations

We'll be using single table inheritance, so this will be used to store the subclass of the TransientRegistration.

We also create and run a database migration for type. If there are any existing TransientRegistrations in the DB when the subtypes are added, they will become NewRegistrations by default.

All TransientRegistrations must have a subtype - for example, NewRegistration.

* Add NewRegistration

Our first subtype of TransientRegistration.

Since we had previously used TransientRegistrations for everything which is now specific to a NewRegistration, this requires a lot of updates.

In our FormsController, when no matching TransientRegistration exists, we create a NewRegistration. If the user enters a form but has no TransientRegistration to work with, we specifically want to create a NewRegistration, not just any type.

This also has a large impact on our testing and factories.

Some of the shared examples are updated to accept multiple types of factory. Other tests just use NewRegistration instead of TransientRegistration.

We also create a shared example to test TransientRegistration behaviour. Since we won't have any non-subclassed TransientRegistrations, this will allow us to test the core behaviour on its subclasses.

This means we can remove transient_reg model specs and factory as these are no longer needed - we don't use the factory for anything, and the test coverage is now provided by the shared example we just created.

* Move new-specific logic from TransientRegistration to NewRegistration

Some parts of the TransientRegistration are actually only needed for the NewRegistration.

Only NewRegistrations will need to generate a new reference. They are also the only subtype that will use the current workflow.

We also rename the workflow concern. This workflow is now specifically the workflow for new registrations, so let's give it a more specific name for clarity's sake.