Implement editing (#139)
https://eaflood.atlassian.net/browse/RUBY-62

This PR adds editing functionality to the engine. It does this by creating a new type of TransientRegistration and a specific editing workflow.

* Add EditRegistration model and workflow

This new model subclasses TransientRegistration and copies data from a Registration when it initialises (similar to how renewals work in waste carriers). It also makes transient copies of the related objects, like addresses, people and exemptions.

We also add a new edit workflow. This works in the same way as the new registration workflow - we have a central state machine managing everything, which is included in the EditRegistration as a concern.

The key difference is that the edit workflow is much less sequential. Generally, users will go from the edit overview form to a specific form, and then back again. In some cases, editing a value requires going through multiple forms/states before returning to the edit overview.

Because the workflow is so much less linear, we end up with a lot more transitions. However, most of the states involved are exactly the same as the other workflow, with the exception of a few edit-specific states.

The workflow has a few new forms which weren't needed for the new registration journey. This PR adds the necessary form classes, controllers, views, routes, etc, following the pattern of our other workflow state form objects.

* Add EditCompletionService

This copies the modified data from the edit_registration back to the registration. It also removes the edit_registration as we no longer require it.

It follows our new service pattern.

The EditCompleteFormsController calls this, which also triggers PaperTrail so we have a record of what was changed.

* Set up edit back and finished paths in EditHelper

These are used to go back out of the edit journey, and to leave it when the journey is complete.

It's intended that host apps can override these with a registration-specific page, which is why we have this unused argument.

* Add a new workflow example for monodirectional transitions

These are transitions where going to the previous or next state will result in the same result. For example, when the workflow is that the user starts on form A, goes to form B, and then returns to form A again.