Update confirmation PDF to reflect the latest layout and design
https://eaflood.atlassian.net/browse/RUBY-223

Change the style and layout of confirmation PDF as per ticket's design direction.
Added gov.uk css style and used to style the document in line with the general website style.
The document is created dynamically and render information in an ad-hoc format depending on the nature
of the business and the site addresses types.