Fail form setup when editing not available
If the transient_registration used to set up a form is an EditRegistration, we now run some checks to see if editing is enabled in the app and for that user.

If not, we redirect to a 404.

This should apply to all forms which use an EditRegistration - however for now we only test this against the edit-specific forms for now, as these are the only specs which currently test using one, as opposed to a NewRegistration.

I created a new concern (PermissionChecks) to handle any user-specific permissions checks, as these are starting to pile up and we have a lot of dodgy patching going on in the back office ApplicationController. My plan is to create a new PermissionChecks concern in the back office which inherits from this one and overrides individual checks as necessary. Non-edit permission checks can be moved over in a separate PR.