Cancel an edit (#142)
https://eaflood.atlassian.net/browse/RUBY-62

It should be possible for users to cancel an in-progress edit. Cancelling should destroy the EditRegistration without saving any of the modified data to the original Registration.

We need two new workflow states here - one to check the user actually wants to cancel, and another to confirm that the cancellation has happened and no changes have been saved. This also means we need a new action on the EditFormsController to get to the 'are you sure you want to cancel?' page.

To handle the cancellation, this PR adds a new EditCancellationService. Which is like the EditCompletionService... except it just destroys the temp object without actually doing anything with the data first.

It also adds a shared example for un-back-able final states. We had an example for 'final' states which users can go back out of, but not ones which are truly the end (in this case, 'journey complete' pages). This commit adds that shared example and refactors existing specs to use it.