Fixes error handling redirect to internale error app
https://eaflood.atlassian.net/browse/RUBY-317

Added configuration for `exceptions_app` in order to correctly redirect rails errors to
our engine ErrorsController, so that the views rendered are the internals styled ones.
Added test coverage on content of the view and correct redirect from real error in requests spec.
Since this is an engine, even if the code usually exist in an `application.rb` or environment file, we have to put it in an initializer instead. For testing purposes, an extra mount
of the app as been added to the dummy test environment app config, since we will only mount the app in `production` environment once merged.
As agreed with Andrew, rendering the views is the main part. Missing paths, restyling and re-wording of the messages in the view will be tackled separately.