Don't skip last email request tests
These tests were being skipped until the error handling bug was fixed: https://github.com/DEFRA/waste-exemptions-engine/pull/145

However, when the skip was removed, it turned out that trying to access these routes when the config disabled them actually resulted in a RoutingError, not a 404. I believe this is because of how the route's constraints are set up.