Add a feature switch and permissions check for editing (#147)
https://eaflood.atlassian.net/browse/RUBY-62

Right now we only want to enable editing in the back office. So we need a switch to enable or disable the feature. Host apps can set this in the engine initializer.

This PR also introduces a permissions check for editing. This is set to always pass by default in the engine but is intended to be overwritten in the host apps when needed.

If the transient_registration used to set up a form is an EditRegistration, we now run some checks to see if editing is enabled in the app and for that user. These transition routes from the edit form don't go through the usual form setup process, so we need to add the edit check for them separately.

If the check fails, we redirect to a 404.

This should apply to all forms which use an EditRegistration - however for now we only test this against the edit-specific forms for now, as these are the only specs which currently test using one, as opposed to a NewRegistration.

I created a new concern (PermissionChecks) to handle any user-specific permissions checks, as these are starting to pile up and we have a lot of dodgy patching going on in the back office ApplicationController. My plan is to create a new PermissionChecks concern in the back office which inherits from this one and overrides individual checks as necessary. Non-edit permission checks can be moved over in a separate PR.

This commit also refactors the boolean checks in the engine config to reduce duplication.