Refactor previous state search on event machine spec helper (#158)
Before we were relying on the existence of a `back` event in order to find the previous valid status of a form. This behaviour was failing with the advent of our first not-flexibly-navigatable form which has no `back` event.
The strategy used in the new version is instead about finding an event transition which goes from a `from` to the current status and returns it's from status.