Fix bug on confirmation PDF generation for lookup addresses
Part of https://eaflood.atlassian.net/browse/RUBY-223

Fix bug on the generation of the PDF when the site address is a normal address with a lookup mode.
Previously I have missed to see that when an address is manually insterted it can be coming from both the
lookup address service or being manually insterted. Tests on the server have identified the issue.