Fix permission checking for edit functionality
https://eaflood.atlassian.net/browse/RUBY-330

The behaviour was due to the fact that the `edit` form was responding to the test url.
Multiple issues were identified:
 - Every time the url was getting hit, a new record in the database was created and not cleaned up, even if no edit permissions are active in the application. The problem was solved by refactoring and checking permissions for all edit forms before any other action is taken.
 - Wheter the token starts with "WEX" the behaviour end error thrown differs. That was due to a temptative to initialize an edit registration with a non-valid WEX number. This have been solved by adding an extra check before initializing for the presence of a registration with the received reference. A not found error is otherwise thrown.
 - The redirect to the error 404 was done wrongly, resulting in the browser hanging on a "you are being redirected" page. The problem was in the status code. Since a `redirect_to`
command already sends a redirect status of 302, the browser gets confused and don't follow the redirect. This have been solved by raising a page not found error and letting the rails application to deal with it in the usual way. Tests have been changed accordingly. Notice that a feature tests done with Capybara would have shown this issue during development, and I suspect the extra `status: 404` used with the redirect_to was caused by the requests spec failing (they would have returned a 302, and should have been tested for that).