Copy over to transient Edit registration only active exemptions
https://eaflood.atlassian.net/browse/RUBY-62 this addresses point 1 of the QA document

This define a new association in the Registration model so that we can get an handy list of only active registration. I have defined the scope manually on the `RegistrationExemption` class rather then include the `CanActivateExemption` concern, since we don't really want that class to be able to deal with changing the status of an exemption (yet).
Thought it might make sense to externalise the list of statuses (for validations purposes and the ability to have the status AASM scopes defined in the model) and leave only the transaction code in the `CanActivateExemption`.

In the meantime this does the job :D