Remove wicked_pdf stlesheet helper (#176)
https://eaflood.atlassian.net/browse/RUBY-223

[PR 130](https://github.com/DEFRA/waste-exemptions-back-office-ta/pull/130) in the WEX back office added the ability to generate a PDF confirmation letter for each submitted registration. It is used by NCCC when needing to send confirmation to fully AD users, or if a user requests a written confirmation.

Because it is generated in the backoffice app, it doesn't face the normal issues of using [Wicked_pdf helpers](https://github.com/mileszs/wicked_pdf#wicked_pdf-helpers) from an engine. So it is using `wicked_pdf_stylesheet_link_tag()`, and was happily doing so until [PR #150](https://github.com/DEFRA/waste-exemptions-engine/pull/150) was merged into the engine.

PR #150 updates the format and content of the registration confirmation PDF and as part of that it added our own version of the `wicked_pdf_stylesheet_link_tag()`. We have to add our own when generating PDF's in an engine because wicked_pdf is unable to find assets.

However doing this means that's what the back office is now using, and so it errors when trying to generate the PDF.

The good news is our own version of the helper still didn't work! So we've had to inline the style till we understand why it didn't (see [PR #168](https://github.com/DEFRA/waste-exemptions-engine/pull/168)) and this means its now redundant.

Hence this change which removes the helper and fixes the back office. We'll no doubt need to return to this issue, but as we're not using this helper currently we've taken the decision to simply remove it.