Remove duplicate partnership? method
Spotted in a the `CertificatePresenter` that we had a `partnership?` however we already have that built into the registration via the `CanHaveRegistrationAttributes`.

As the `CertificatePresenter` delegates any calls it does not recognise down to the registration instance, deleting the method will have no impact on anything that uses the presenter.