Fix `cannot edit` label in caption to float right in line with title
https://eaflood.atlassian.net/browse/RUBY-367

This fixes once and for all the position of the `cannot edit` label to be the far right of the caption heading. A `small` tag have been used as it is one of the few W3C allowed tags to nest, however classes have been used to achieve a reduced font size and the floating right behaviour.