Require wicked pdf at a Rails level (#193)
From: https://eaflood.atlassian.net/browse/RUBY-223

This was an hard catch. It all started from a bug arises from RUBY-223 for which the back office application was generating PDFs containing PDFs content.
It wasn't clear what was causing that. We also noticed other bugs for which the wicked_pdf Rails helper for loading static assets to the view were not loaded by the Rails views.
The `wicked_pdf` rerquired gem in the bundler of the back office application was never the issue. It was actually allowing the application to load the gem correctly. And that was causing the `rernder_to_string` ActiveView method to render the content already parsed correctly by the wicked_pdf Rails integration. Hence the GenerratePdfService was wrapping PDF content into a PDF file and causing the issue with the initial bug. This PR gets rid of this code. I have searched the other applications to see if they refer to the class directly but I have found no match.