Do not save un-persisted objects in workflow setup (#188)
* Do not save unpersisted objects in workflow setup

https://eaflood.atlassian.net/browse/RUBY-352

The registration was saved at every refresh because of the workflow setup.
Since the object is persisted already at every submit, I have added a guard clause to avoid updating it unless it is persisted.
BUT the editing journey is instead expecting a persisted object since the first call. Hence a fix there was necessary too.

I wasn't able to find a nice way to test it :/ we should run some tests in truncation mode or do some weird stubbing and expectations.