Add area attribute to addresses
https://eaflood.atlassian.net/browse/RUBY-25

The existing service has the ability to lookup the EA admin area for a site using the easting and northing values. This is currently out of action whilst we await a fix (the external service we call changed supplier and we have been awaiting support for what the new query should be).

However it has been spotted in our work on migrating the old data to the new schema that we don't have a field for this. So on the basis we want to retain this data, and that we expect to implement the lookup feature in the new service once the issue is resolved, this change adds a migration to add the required field.