Auto-determine grid reference for site address
https://eaflood.atlassian.net/browse/RUBY-25

Whilst working on the migration of data from the old schema to the new schema we identified (remembered!) that the existing service automatically determines the grid reference if an address is used for the site.

This information is then what gets exported to the Electronic Public Register, and is used in its 'site is within x miles of postcode' search.

We will currently attempt to determine the x & y for manually entered site addresses in order to support the EA admin area lookup (to be implemented), but we now know we also need to be determining the grid reference.

This change updates the `create_from_x()` methods in the `transient_address` model to also include an attempt to determine the grid reference for a site.