Add tests for auto x & y population
Was intending to add tests for the auto grid reference population, but then spotted we don't seem to be covering the existing auto x & y population.

So added this test to cover ensuring we look for and populate if found the x & y values for a site address if the source was a manual address entry.