Add versions archive table migration (#218)
https://eaflood.atlassian.net/browse/RUBY-25

As part of the first release of this version of the service we will be migrating data from the existing version into our new schema.

The migration will include audit history stored in the existing data. We already have a migration in place and tested for the user records. As the structure of the user records has changed very little this has not been an issue, and how [papertrail](https://github.com/paper-trail-gem/paper_trail) is working for the User model is sufficient.

However when it comes to the registrations, the structure has changed considerably. Plus we reviewed how auditing is currently working in this version of the service to help inform the migration work. What we found is that

- when we create a registration we get 4 entries; 1 registration `create` and 3 address `create` entries
- if we edit the registration in the back office we get; 1 registration `update` and 3 address `create` entries
- if we edit only an address on the registration (no direct registration properties) we get; just 3 address `create` entries

This is because we have paper trail attached to the address model (taking our lead from the current service, but the way this version works is that addresses get recreated each time a change occurs. So there will never be an edit.

So this is not reflecting what we feel is actually taking place, nor that we consider the addresses as being part of the registration, not something separate. As such we are looking at how we can improve our audit functionality.

This is a long winded way of saying

- the existing audit history data does not match with how this version works
- the existing audit history data will definately not match with how we are likely to make auditing work
- the existing audit history data isn't that great anyway

As such the migration will move the data to an archive table, which we can access if ever needed but which to begin with we are not intending to represent in a model or linked to the `Registration`.

This means we need a new table, hence this change to create it using active record migrations, so that all structure and logic related to the database remains in one place.