Store a single version of registration and its relations
When a registration's address changes, or some other object belonging to the registration changes,  we want to store that the registration was updated.

To accomplish this, we've decided to store tracking addresses alone and instead force a new version to be saved for the registration when an edit is completed. We'll also store the "full" registration data with all the related object data as JSON so we can get a full record.