Save a version when editing even if only related object has changed
Addresses and people will always be recreated, but this checks to see if their attributes have changed or not. We only save a version if the information has actually changed.

However, the tests highlighted a new bug - we are saving the JSON for the updated object, not the last version. So next step is to fix that.