Fix related object comparison
Previously our checks to see if relations had changed were always passing, because the order of the relations had changed, even if the core data didn't. Sorting the relations before comparing them resulted in the comparison always failing, because we had already updated the data and no longer had the old record to compare it to.

So instead we perform the comparison before we do any updates, and then use the result later to decide if we need to save a new version or not.