Store a single version of registration and its relations (#220)
When a registration's address changes, or some other object belonging to the registration changes,  we want to store that the registration was updated.

To accomplish this, we've decided to stop tracking addresses alone and instead force a new version to be saved for the registration when an edit is completed. We'll also store the "full" registration data with all the related object data as JSON so we can get a full record.

We only save a version when editing if the data has actually changed. Addresses and people will always be recreated, but this checks to see if the actual attribute values have changed or not.